+++ lib/app/router/router.dart
@@ -1,6 +1,6 @@
 import 'package:go_router/go_router.dart';
 import 'package:logistic_operation/features/auth/presentaion/pages/login_page.dart';
-import 'package:logistic_operation/features/dashboard/presentation/pages/dashboard.dart';
+import 'package:logistic_operation/features/dashboard/presentation/pages/dashboard_page.dart';
 import 'package:logistic_operation/features/splash/presentation/pages/splash_page.dart';
 
 class AppRouter {