+++ lib/features/auth/presentaion/providers/auth_provider.dart
@@ -1,7 +1,9 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:logistic_operation/core/providers/storage_provider.dart';
 import 'package:logistic_operation/features/auth/data/repository/auth_repository.dart';
+import 'package:logistic_operation/features/auth/domain/usecases/check_login_usecase.dart';
 import 'package:logistic_operation/features/auth/domain/usecases/login_usecase.dart';
+import 'package:logistic_operation/features/auth/domain/usecases/logout_usecase.dart';
 
 import '../../data/datasource/auth_remote_datasource.dart';
 import '../../data/datasource/auth_remote_datasource_impl.dart';
@@ -20,4 +22,12 @@ final authRepositoryProvider = Provider<AuthRepository>((ref) {
 
 final loginUseCaseProvider = Provider<LoginUseCase>((ref) {
   return LoginUseCase(ref.read(authRepositoryProvider));
+});
+
+final checkLoginUseCaseProvider = Provider<CheckLoginUseCase>((ref) {
+  return CheckLoginUseCase(ref.read(authRepositoryProvider));
+});
+
+final logoutUseCaseProvider = Provider<LogoutUseCase>((ref) {
+  return LogoutUseCase(ref.read(authRepositoryProvider));
 });