+++ lib/features/dashboard/domain/repository/dashboard_repository.dart
@@ -1,0 +1,5 @@
+import '../entities/dashboard_summary.dart';
+
+abstract interface class DashboardRepository {
+  Future<DashboardSummary> getSummary();
+}