+++ lib/features/dashboard/domain/usecases/get_dashboard_summary_usecase.dart
@@ -1,0 +1,12 @@
+import '../entities/dashboard_summary.dart';
+import '../repository/dashboard_repository.dart';
+
+class GetDashboardSummaryUseCase {
+  final DashboardRepository repository;
+
+  const GetDashboardSummaryUseCase(this.repository);
+
+  Future<DashboardSummary> call() {
+    return repository.getSummary();
+  }
+}