+++ lib/features/logistics/shipment/domain/repository/shipment_repository.dart
@@ -1,0 +1,5 @@
+import '../entities/shipment.dart';
+
+abstract interface class ShipmentRepository {
+  Future<List<Shipment>> getRecentShipments();
+}