+++ lib/features/logistics/shipment/domain/usecases/get_recent_shipment_usecase.dart
@@ -1,0 +1,12 @@
+import '../entities/shipment.dart';
+import '../repository/shipment_repository.dart';
+
+class GetRecentShipmentsUseCase {
+  final ShipmentRepository repository;
+
+  const GetRecentShipmentsUseCase(this.repository);
+
+  Future<List<Shipment>> call() {
+    return repository.getRecentShipments();
+  }
+}