+++ lib/features/splash/presentation/pages/splash_page.dart
@@ -1,7 +1,37 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:go_router/go_router.dart';
+import 'package:logistic_operation/features/auth/presentaion/providers/auth_notifier.dart';
 
-class SplashPage extends StatelessWidget {
+class SplashPage extends ConsumerStatefulWidget {
   const SplashPage({super.key});
+
+  @override
+  ConsumerState<SplashPage> createState() => _SplashPageState();
+}
+
+class _SplashPageState extends ConsumerState<SplashPage> {
+  @override
+  void initState() {
+    super.initState();
+    _initialize();
+  }
+
+  Future<void> _initialize() async {
+    await Future.delayed(const Duration(seconds: 1));
+
+    final isLoggedIn = await ref
+        .read(authNotifierProvider.notifier)
+        .checkLogin();
+
+    if (!mounted) return;
+
+    if (isLoggedIn) {
+      context.go('/dashboard');
+    } else {
+      context.go('/login');
+    }
+  }
 
   @override
   Widget build(BuildContext context) {